add state machine

+++ common/native/sys_time/include/sys_time/ClockSource.hpp
@@ -15,19 +15,19 @@ struct ClockSource
    static uint32_t now_us() noexcept
    {
       using namespace std::chrono;
-      return duration_cast<microseconds>(steady_clock::now().time_since_epoch()).count();
+      return static_cast<uint32_t>(duration_cast<microseconds>(steady_clock::now().time_since_epoch()).count());
    }
 
    static uint32_t now_ms() noexcept
    {
       using namespace std::chrono;
-      return duration_cast<milliseconds>(steady_clock::now().time_since_epoch()).count();
+      return static_cast<uint32_t>(duration_cast<milliseconds>(steady_clock::now().time_since_epoch()).count());
    }
 
    static uint32_t now_s() noexcept
    {
       using namespace std::chrono;
-      return duration_cast<seconds>(steady_clock::now().time_since_epoch()).count();
+      return static_cast<uint32_t>(duration_cast<seconds>(steady_clock::now().time_since_epoch()).count());
    }
 };
